Fix No arguments, expected User(s), Forum(s) or Topic(s) in .../app/Models/Topic/Delete.php:[29]

+++ app/Models/Topic/Delete.php
@@ -115,7 +115,9 @@ class Delete extends Action
 //                $parents[$topic->parent->id] = $topic->parent;
 //            }
 
-            $this->delete(...($this->manager->loadByIds($tids, false)));
+            if ($tids) {
+                $this->delete(...($this->manager->loadByIds($tids, false)));
+            }
         }
 
         $this->c->posts->delete(...$args);
